feature: 로또 번호는 1부터 45까지의 숫자이다.
