+++ src/main/java/step2/util/GeneratorUtil.java
@@ -1,4 +1,4 @@
-package step3.util;
+package step2.util;
 
 import java.util.ArrayList;
 import java.util.Collections;
