+++ src/main/java/step2/view/InputView.java
@@ -1,4 +1,4 @@
-package step3.view;
+package step2.view;
 
 import java.util.Arrays;
 import java.util.List;
