+++ src/test/java/step2/domain/WinningLottoTest.java
@@ -1,4 +1,4 @@
-package step3.domain;
+package step2.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
